Fix saving order changes
Remove the the add-item class from the Save Changes button to prevent
the click handler that shows the add_inventory_modal dialog from being
attached to the button. When the Save Changes button has the add-item
class, the click handler causes the Add Inventory Item dialog to pop
up when the button is pressed.